requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.2.0
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2
django-import-export==3.3.7
django-redis==5.4.0
dnspython==2.6.1
et-xmlfile==1.1.0
eventlet==0.35.2
flower==2.0.1
greenlet==3.0.3
humanize==4.9.0
idna==3.6
kombu==5.3.5
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.1.2
prometheus_client==0.20.0
prompt-toolkit==3.0.43
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
redis==5.0.2
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
tablib==3.5.0
tornado==6.4
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
wget==3.2
xlrd==2.0.1
xlwt==1.3.0
